- refactored experiment model to merge v1 with v2
