+++ api/turing/validation/validator_test.go
@@ -16,8 +16,8 @@ import (
 	"github.com/gojek/turing/api/turing/models"
 	"github.com/gojek/turing/api/turing/service/mocks"
 	"github.com/gojek/turing/api/turing/validation"
-	"github.com/gojek/turing/engines/experiment/v2"
-	"github.com/gojek/turing/engines/experiment/v2/manager"
+	"github.com/gojek/turing/engines/experiment"
+	"github.com/gojek/turing/engines/experiment/manager"
 )
 
 func TestValidateLogConfig(t *testing.T) {
@@ -244,7 +244,7 @@ func TestValidateTrafficRules(t *testing.T) {
 				{
 					Conditions: []*router.TrafficRuleCondition{
 						{
-							FieldSource: experiments.HeaderFieldSource,
+							FieldSource: experiment.HeaderFieldSource,
 							Field:       "X-Region",
 							Operator:    router.InConditionOperator,
 							Values:      []string{"region-a", "region-b"},
@@ -293,7 +293,7 @@ func TestValidateTrafficRules(t *testing.T) {
 				{
 					Conditions: []*router.TrafficRuleCondition{
 						{
-							FieldSource: experiments.HeaderFieldSource,
+							FieldSource: experiment.HeaderFieldSource,
 							Field:       "X-Region",
 							Operator:    router.InConditionOperator,
 							Values:      []string{"region-b"},
@@ -325,7 +325,7 @@ func TestValidateTrafficRules(t *testing.T) {
 				{
 					Conditions: []*router.TrafficRuleCondition{
 						{
-							FieldSource: experiments.HeaderFieldSource,
+							FieldSource: experiment.HeaderFieldSource,
 							Field:       "X-Region",
 							Operator:    router.RuleConditionOperator{},
 							Values:      []string{"region-b"},
@@ -389,7 +389,7 @@ func TestValidateTrafficRules(t *testing.T) {
 				{
 					Conditions: []*router.TrafficRuleCondition{
 						{
-							FieldSource: experiments.HeaderFieldSource,
+							FieldSource: experiment.HeaderFieldSource,
 							Field:       "",
 							Operator:    router.InConditionOperator,
 							Values:      []string{},
@@ -417,7 +417,7 @@ func TestValidateTrafficRules(t *testing.T) {
 				{
 					Conditions: []*router.TrafficRuleCondition{
 						{
-							FieldSource: experiments.PayloadFieldSource,
+							FieldSource: experiment.PayloadFieldSource,
 							Field:       "some_property",
 							Operator:    router.InConditionOperator,
 							Values:      []string{"some_value"},
@@ -443,7 +443,7 @@ func TestValidateTrafficRules(t *testing.T) {
 				{
 					Conditions: []*router.TrafficRuleCondition{
 						{
-							FieldSource: experiments.PayloadFieldSource,
+							FieldSource: experiment.PayloadFieldSource,
 							Field:       "some_property",
 							Operator:    router.InConditionOperator,
 							Values:      []string{"some_value"},
