+++ engines/experiment/go.mod
@@ -3,8 +3,12 @@ module github.com/gojek/turing/engines/experiment
 go 1.14
 
 require (
-	github.com/gojek/turing/engines/experiment/v2 v2.0.0
+	github.com/buger/jsonparser v1.1.1
+	github.com/go-playground/universal-translator v0.18.0 // indirect
+	github.com/go-playground/validator v9.31.0+incompatible
+	github.com/hashicorp/go-plugin v1.4.3
+	github.com/leodido/go-urn v1.2.1 // indirect
+	github.com/pkg/errors v0.9.1
 	github.com/stretchr/testify v1.6.1
+	gopkg.in/go-playground/assert.v1 v1.2.1 // indirect
 )
-
-replace github.com/gojek/turing/engines/experiment/v2 => ../experiments/
