+++ engines/experiment/manager/mocks/ExperimentManager.go
@@ -3,7 +3,7 @@
 package mocks
 
 import (
-	"github.com/gojek/turing/engines/experiment/v2/manager"
+	"github.com/gojek/turing/engines/experiment/manager"
 	mock "github.com/stretchr/testify/mock"
 )
