+++ engines/experiment/plugin.go
@@ -1,4 +1,4 @@
-package experiments
+package experiment
 
 import "github.com/hashicorp/go-plugin"
