+++ engines/experiment/request.go
@@ -1,6 +1,7 @@
-package experiments
+package experiment
 
 import (
+	"encoding/json"
 	"fmt"
 	"net/http"
 	"strings"
@@ -9,6 +10,40 @@ import (
 	"github.com/buger/jsonparser"
 	"github.com/pkg/errors"
 )
+
+// FieldSource is used to identify the source of the Litmus user data field
+type FieldSource string
+
+const (
+	// PayloadFieldSource is used to represent the request payload
+	PayloadFieldSource FieldSource = "payload"
+	// HeaderFieldSource is used to represent the request header
+	HeaderFieldSource FieldSource = "header"
+)
+
+// GetFieldSource converts the input string to a FieldSource
+func GetFieldSource(srcString string) (FieldSource, error) {
+	switch strings.ToLower(srcString) {
+	case "header":
+		return HeaderFieldSource, nil
+	case "payload":
+		return PayloadFieldSource, nil
+	}
+	return FieldSource(""), fmt.Errorf("Unknown field source %s", srcString)
+}
+
+// UnmarshalJSON unmarshals the data as a string and then creates the
+// appropriate FieldSource
+func (f *FieldSource) UnmarshalJSON(data []byte) error {
+	var s string
+	var err error
+	if err = json.Unmarshal(data, &s); err != nil {
+		return err
+	}
+
+	*f, err = GetFieldSource(s)
+	return err
+}
 
 // GetValueFromRequest parses the request header / payload to retrieve the value
 // for the given field
