+++ engines/router/missionctl/experiment/experiment_test.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"testing"
 
-	_ "github.com/gojek/turing/engines/experiment/runner/nop"
+	_ "github.com/gojek/turing/engines/experiment/v1/runner/nop"
 	"github.com/gojek/turing/engines/router/missionctl/turingctx"
 	"github.com/stretchr/testify/assert"
 )
