+++ engines/router/missionctl/fiberapi/fan_in_test.go
@@ -15,6 +15,7 @@ import (
 	"github.com/gojek/fiber"
 	fiberhttp "github.com/gojek/fiber/http"
 	"github.com/gojek/turing/engines/experiment/runner"
+	runnerV1 "github.com/gojek/turing/engines/experiment/v1/runner"
 	"github.com/gojek/turing/engines/router/missionctl/experiment"
 	tfu "github.com/gojek/turing/engines/router/missionctl/fiberapi/internal/testutils"
 	tu "github.com/gojek/turing/engines/router/missionctl/internal/testutils"
@@ -95,7 +96,7 @@ func TestInitializeEnsemblingFanIn(t *testing.T) {
 
 			// Monkey patch functionality that is external to the current package and run
 			monkey.Patch(
-				runner.Get,
+				runnerV1.Get,
 				func(name string, config json.RawMessage) (runner.ExperimentRunner, error) {
 					return nil, nil
 				},
@@ -109,7 +110,7 @@ func TestInitializeEnsemblingFanIn(t *testing.T) {
 			err := fanIn.Initialize(data.properties)
 
 			monkey.Unpatch(experiment.NewExperimentRunner)
-			monkey.Unpatch(runner.Get)
+			monkey.Unpatch(runnerV1.Get)
 
 			// Test error and if fanIn is initialised as expected
 			assert.Equal(t, data.success, err == nil)
