+++ engines/router/missionctl/fiberapi/routing_policies_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"github.com/gojek/turing/engines/experiment/runner/nop"
+	_ "github.com/gojek/turing/engines/experiment/v1/runner/nop"
 	tu "github.com/gojek/turing/engines/router/missionctl/internal/testutils"
 	"github.com/stretchr/testify/assert"
 )
