+++ engines/router/missionctl/fiberapi/traffic_splitting_strategy_test.go
@@ -15,7 +15,7 @@ import (
 	tu "github.com/gojek/turing/engines/router/missionctl/internal/testutils"
 	"github.com/gojek/turing/engines/router/missionctl/turingctx"
 
-	"github.com/gojek/turing/engines/experiment/v2"
+	"github.com/gojek/turing/engines/experiment"
 	"github.com/stretchr/testify/require"
 )
 
@@ -33,13 +33,13 @@ func TestTrafficSplittingStrategyRule_TestRequest(t *testing.T) {
 			rule: &fiberapi.TrafficSplittingStrategyRule{
 				Conditions: []*router.TrafficRuleCondition{
 					{
-						FieldSource: experiments.HeaderFieldSource,
+						FieldSource: experiment.HeaderFieldSource,
 						Field:       "X-Region",
 						Operator:    router.InConditionOperator,
 						Values:      []string{"region-a", "region-b"},
 					},
 					{
-						FieldSource: experiments.PayloadFieldSource,
+						FieldSource: experiment.PayloadFieldSource,
 						Field:       "foo.bar",
 						Operator:    router.InConditionOperator,
 						Values:      []string{"foobar"},
@@ -56,19 +56,19 @@ func TestTrafficSplittingStrategyRule_TestRequest(t *testing.T) {
 			rule: &fiberapi.TrafficSplittingStrategyRule{
 				Conditions: []*router.TrafficRuleCondition{
 					{
-						FieldSource: experiments.HeaderFieldSource,
+						FieldSource: experiment.HeaderFieldSource,
 						Field:       "X-Region",
 						Operator:    router.InConditionOperator,
 						Values:      []string{"region-a", "region-b"},
 					},
 					{
-						FieldSource: experiments.PayloadFieldSource,
+						FieldSource: experiment.PayloadFieldSource,
 						Field:       "another",
 						Operator:    router.InConditionOperator,
 						Values:      []string{"42"},
 					},
 					{
-						FieldSource: experiments.PayloadFieldSource,
+						FieldSource: experiment.PayloadFieldSource,
 						Field:       "foo.bar",
 						Operator:    router.InConditionOperator,
 						Values:      []string{"foobar"},
@@ -85,19 +85,19 @@ func TestTrafficSplittingStrategyRule_TestRequest(t *testing.T) {
 			rule: &fiberapi.TrafficSplittingStrategyRule{
 				Conditions: []*router.TrafficRuleCondition{
 					{
-						FieldSource: experiments.HeaderFieldSource,
+						FieldSource: experiment.HeaderFieldSource,
 						Field:       "X-Region",
 						Operator:    router.InConditionOperator,
 						Values:      []string{"region-a", "region-b"},
 					},
 					{
-						FieldSource: experiments.PayloadFieldSource,
+						FieldSource: experiment.PayloadFieldSource,
 						Field:       "another",
 						Operator:    router.InConditionOperator,
 						Values:      []string{"42"},
 					},
 					{
-						FieldSource: experiments.PayloadFieldSource,
+						FieldSource: experiment.PayloadFieldSource,
 						Field:       "foo.bar",
 						Operator:    router.InConditionOperator,
 						Values:      []string{"foobar"},
@@ -163,7 +163,7 @@ func TestTuringTrafficSplittingStrategy_Initialize(t *testing.T) {
 						RouteID: "ROUTE-A",
 						Conditions: []*router.TrafficRuleCondition{
 							{
-								FieldSource: experiments.HeaderFieldSource,
+								FieldSource: experiment.HeaderFieldSource,
 								Field:       "Content-Type",
 								Operator:    router.InConditionOperator,
 								Values:      []string{"application/json"},
@@ -174,7 +174,7 @@ func TestTuringTrafficSplittingStrategy_Initialize(t *testing.T) {
 						RouteID: "ROUTE-B",
 						Conditions: []*router.TrafficRuleCondition{
 							{
-								FieldSource: experiments.HeaderFieldSource,
+								FieldSource: experiment.HeaderFieldSource,
 								Field:       "Content-Type",
 								Operator:    router.InConditionOperator,
 								Values:      []string{"application/xml"},
@@ -230,13 +230,13 @@ func TestTrafficSplittingStrategy_SelectRoute(t *testing.T) {
 						RouteID: "route-a",
 						Conditions: []*router.TrafficRuleCondition{
 							{
-								FieldSource: experiments.HeaderFieldSource,
+								FieldSource: experiment.HeaderFieldSource,
 								Field:       "X-Region",
 								Operator:    router.InConditionOperator,
 								Values:      []string{"region-a", "region-b"},
 							},
 							{
-								FieldSource: experiments.PayloadFieldSource,
+								FieldSource: experiment.PayloadFieldSource,
 								Field:       "service_type",
 								Operator:    router.InConditionOperator,
 								Values:      []string{"service-b"},
@@ -263,13 +263,13 @@ func TestTrafficSplittingStrategy_SelectRoute(t *testing.T) {
 						RouteID: "route-a",
 						Conditions: []*router.TrafficRuleCondition{
 							{
-								FieldSource: experiments.HeaderFieldSource,
+								FieldSource: experiment.HeaderFieldSource,
 								Field:       "X-Region",
 								Operator:    router.InConditionOperator,
 								Values:      []string{"region-a", "region-b"},
 							},
 							{
-								FieldSource: experiments.PayloadFieldSource,
+								FieldSource: experiment.PayloadFieldSource,
 								Field:       "service_type",
 								Operator:    router.InConditionOperator,
 								Values:      []string{"service-b"},
@@ -280,7 +280,7 @@ func TestTrafficSplittingStrategy_SelectRoute(t *testing.T) {
 						RouteID: "route-b",
 						Conditions: []*router.TrafficRuleCondition{
 							{
-								FieldSource: experiments.PayloadFieldSource,
+								FieldSource: experiment.PayloadFieldSource,
 								Field:       "service_type",
 								Operator:    router.InConditionOperator,
 								Values:      []string{"service-a", "service-b"},
@@ -305,13 +305,13 @@ func TestTrafficSplittingStrategy_SelectRoute(t *testing.T) {
 						RouteID: "route-a",
 						Conditions: []*router.TrafficRuleCondition{
 							{
-								FieldSource: experiments.HeaderFieldSource,
+								FieldSource: experiment.HeaderFieldSource,
 								Field:       "X-Region",
 								Operator:    router.InConditionOperator,
 								Values:      []string{"region-a", "region-b"},
 							},
 							{
-								FieldSource: experiments.PayloadFieldSource,
+								FieldSource: experiment.PayloadFieldSource,
 								Field:       "service_type",
 								Operator:    router.InConditionOperator,
 								Values:      []string{"service-b"},
@@ -322,7 +322,7 @@ func TestTrafficSplittingStrategy_SelectRoute(t *testing.T) {
 						RouteID: "route-b",
 						Conditions: []*router.TrafficRuleCondition{
 							{
-								FieldSource: experiments.PayloadFieldSource,
+								FieldSource: experiment.PayloadFieldSource,
 								Field:       "service_type",
 								Operator:    router.InConditionOperator,
 								Values:      []string{"service-a", "service-b"},
@@ -349,7 +349,7 @@ func TestTrafficSplittingStrategy_SelectRoute(t *testing.T) {
 						RouteID: "route-a",
 						Conditions: []*router.TrafficRuleCondition{
 							{
-								FieldSource: experiments.HeaderFieldSource,
+								FieldSource: experiment.HeaderFieldSource,
 								Field:       "X-Region",
 								Operator:    router.InConditionOperator,
 								Values:      []string{"region-a"},
@@ -375,7 +375,7 @@ func TestTrafficSplittingStrategy_SelectRoute(t *testing.T) {
 						RouteID: "route-a",
 						Conditions: []*router.TrafficRuleCondition{
 							{
-								FieldSource: experiments.HeaderFieldSource,
+								FieldSource: experiment.HeaderFieldSource,
 								Field:       "X-Region",
 								Operator:    router.InConditionOperator,
 								Values:      []string{"region-a"},
