+++ engines/router/missionctl/handlers/batch_http_handler_test.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 	"time"
 
-	_ "github.com/gojek/turing/engines/experiment/runner/nop"
+	_ "github.com/gojek/turing/engines/experiment/v1/runner/nop"
 	"github.com/gojek/turing/engines/router/missionctl"
 	"github.com/gojek/turing/engines/router/missionctl/config"
 	"github.com/stretchr/testify/assert"
