+++ engines/router/missionctl/mission_control_test.go
@@ -17,7 +17,7 @@ import (
 	"testing"
 	"time"
 
-	_ "github.com/gojek/turing/engines/experiment/runner/nop"
+	_ "github.com/gojek/turing/engines/experiment/v1/runner/nop"
 	"github.com/gojek/turing/engines/router/missionctl/config"
 	"github.com/gojek/turing/engines/router/missionctl/errors"
 	"github.com/gojek/turing/engines/router/missionctl/fiberapi"
